This is going to be hard
It seems we are currently trying to run Chrome on Travis. Which is a
bit hard (https://github.com/travis-ci/travis-ci/issues/938). Firefox
is recommended:
https://docs.travis-ci.com/user/gui-and-headless-browsers/#Using-xvfb-to
-Run-Tests-That-Require-a-GUI

But it's still only Firefox.

We can keep the chrome work for local tests, but I think Travis should
ping Sauce Labs:
https://docs.travis-ci.com/user/gui-and-headless-browsers/#Using-Sauce-L
abs

So they can spin up 180 real browsers and test on those.