Prevent duplicate state update when setting thumbnail
Now that thumbnail generation is synchronous using URL.createObjectURL,
we can set the thumbnail URL on the file object immediately. This saves
a state update, since `addThumbnail` previously would also trigger a
state update. That used to make sense, because thumbnails were generated
async using a FileReader, but it's not necessary anymore.