webcam,dashboard: Add `isSupported()` API for providers
Providers can signal whether they support the current environment using
an optional `isSupported()` method. Providers that always work (eg.
Google Drive) do not need to specify this method.

The check for the Webcam currently only checks if the `getUserMedia`
method exists at all. We could maybe do some more checks, like, if the
mode is `video-only` or `video-audio` but MediaRecorder is not
supported.