Modify file.name to return just the filename in IE
In IE `file.name` returns the full path to the user's file on the disk.
We will never need the path to the file; even if we for some reason
wanted to, that couldn't be cross-browser because only IE gives us that
information.

So we update `file.name` to remove the directories and leave only the
filename.