s3: Fix uploadURL for presigned PUT uploads
This strips the query string from the URL used for a successful PUT
upload to determine the `uploadURL`. For PUT uploads, the path that you
use for the upload is the same as the path you use to access the object.
But the query string contains a signature for a PUT request, so it can't
be used for a GET request. Simply removing it is not always enough (the
bucket may restrict GET requests) but it's better than having a PUT
signature in there.

Ref https://github.com/transloadit/uppy/issues/815