transloadit: Only use socket.io's WebSocket transport. (#1029)
There is not much benefit to this (except saving a few HTTP requests)
but if we do this we'll see if anyone's firewalls are blocking WS or
something. Then one day in the future we can switch to plain WebSockets
instead of socket.io to shed 20KB once the Transloadit backend supports
it.