transloadit: throw and warn when using the old Uppy Server URL (#1038)
Shows up in the informer and in the console when trying to upload a file
from a remote source. Kinda sucks that you have to actually do it, but
this is the easiest place to implement it…else we have to look at other
plugins' private options and hope that users are adding the plugins in
an order that makes that possible.