companion: drop parallel down/upload for S3 multipart
We were not ending the `fs-tail-stream` correctly. `fs-tail-stream` is
also not a battle-tested package and relying on file watching is never a
fun time, so it's better if we can avoid it entirely.

One good way to do that is to axe the entire parallel download/upload
feature, and instead download files in full before uploading them, like
we do for Tus and XHR :)

Fixes #1457.