screen-capture: add type tests
plugins have to be explicitly excluded in `npm run test:type` if they
don't have a type test file, so this was causing tests to fail.

updated the types to the new style capable of strict typechecking, and
made them use the actual DOM types for the Constraints options.

Looking at mdn, the `displaySurface` and similar options are actually
properties of the `constraints.video` object, so i moved them in there
in our `defaultOptions`.