website: list robodog size, sort plugins by name
when looking for the size of a plugin, I could never find it without
Ctrl+F because the order was essentially arbitrary. This change sorts
the plugins in the package size table alphabetically, while still
keeping the `uppy` and `@uppy/core` packages on top, and the
`@uppy/store-` packages at the bottom (unlikely to be interesting to
most people).