Fix: Incorrect File Added Count
On the dashboard, when uploading a folder, if files from that folder are not successfully added, the count would still say they are there

For example, If you're using Google Drive, and upload a folder with 3 images and 2 mp3 files. If you also have a restriction set to only allow images, Dashboard would still read something like "5 Files Added", even though only 3 would be there. I resolved this by making the `addFile` method in the `provider-views` package return a boolean value based on whether or not the file was added successfully. I then created a count variable with the `addFolder` method and added to that count for every successful addition. Using that count instead of the total count of files in the folder fixes this problem

Resolves #2038