dashboard: remove preact-css-transition-group
Unfortunately I can't figure out how to make the dependencies and
versions work out well here.

Instead, this uses a custom slide transition component. It is much more
simple than preact-css-transition-group, which supports animating groups
of components. The `<Slide />` component:

- can only support one element, that either exists or does not exist
- requires that the child element takes a `className` property, that it
  passes onto its root element
- only uses `setTimeout()` for when the transition finishes
  (in preact-css-transition-group, the `transitionend` event is used,
  which is more _accurate_ but less _reliable_)
- hardcodes the transition duration
- hardcodes the transition name

This makes it much more limited. Hopefully the reduced complexity is
worth it.

If we have other transitions in the future, we can copy-paste this
component.