dashboard: fix showing showProgressDetails on md and up
The DOM changed to accomodate RTL scripts, but this CSS rule was not
changed accordingly.

The StatusBar does not know its own width, so a standalone StatusBar
would never have an `.uppy-size--md` parent element. You only get that
with the Dashboard. So I've moved the rule to the Dashboard plugin, and
the selector now checks for `.uppy-Dashboard`. That element also has the
size classes.

The standalone StatusBar does not auto-hide progress details on narrow
screens. I think that's probably fine, if you're using it directly and
explicitly asking for progress details you might want to always see
them…