companion: support relative redirect URLs in responses
When a URL import attempt responds with a redirect, we tried to naively
parse the URL with `new URL()`. This is a problem when the URL is
invalid, because that constructor will throw an error.

Redirect URLs are also allowed to be relative URLs, which was not
supported by us.

With this change, invalid URLs will cause a redirect to be blocked,
and relative redirects will work.