core: tighten duck type check for file objects
When the `'error'` event is emitted, the error is stored on the file
object if the second argument is a file.

I don't know if that is intentional behaviour or if it's a coincidental
effect of the implementation. We don't use it anywhere.

When an upload step fails, `Uppy#_runUpload()` may emit an 'error'
message with the upload ID in the second argument. So then `file` is the
upload ID in this event handler. A string is not null, but it also
doesn't have an `.id` property, so the `setFileState()` call would fail.

Now it checks to make sure that the second argument actually could be a
file.