remove google drive link
https://app.intercom.com/a/apps/qiqpfgjg/inbox/inbox/all/conversations/26852700005169 
No. 3i, removed the link as I couldn't find a replacement. The only link I could find that seemed to be about the same thing as the original link is: https://github.com/transloadit/uppy/blob/main/website/src/docs/google-drive.md, but this is already available on the website for users to see.