Update Url.jsx
Uppy Url is great! Thank you. For my use case, I need to be able to pass thru some meta data with the File, in addition to the URL.  Harmless to make it optional, but very useful to have it.  Without it, I had to copy/paste the code in addFile to add it externally, since there's no intermediate opportunity to add data before uppy.addFile gets called.