transloadit: propagate error details when creating Assembly fails
Clients are expected to use `error.assembly` to inspect details about
failed Assemblies, but then we do need to always give that property
back.

I think we should have a test for this but not sure how to approach it,
since it depends on responses from the Transloadit API. I guess we'd
require a new e2e test for it?