meta: deploy to Heroku on every companion commit (#4367)
Instead of deploying on each release, deploy to Heroku on every commit that is touching a companion file (except release commits, but that's fine as the release commit should contain doc-only changes). This is to increase our chance of catching bugs before releases, and would ensure that Uppy.io examples are using the edge.