fix some types in core
1. don't type assert opts, but instead use default values
so that typescript understands it.
This will prevent `this.opts` looking like they are required
when they in reality are undefined (at runtime).
This change also uncovered the fact that `locale` is in fact optional,
so change types to reflect the fact.
This change also fixes the problem where any options set to `undefined` in the constructor
would overwrite any default options, example:
{ ...{ id: 'uppy' }, ...{ id: undefined } } becomes { id: undefined }

2. remove some other unsafe type assertions