fix: derive unique encryption keys
Closes #5706

Using the same encryption key to encrypt an unlimited amount of
data/documents results in key wear-out, in particular when using AES-CBC
or other cipher modes that are not very resistant to nonce misuse.

This commit derives from the application secret a fresh encryption key
and a fresh initialization vector for every document to encrypt. HKDF
is used to derive these values and the context info parameter is set to
a fresh random value called nonce. This nonce replaces the IV in the
token.

Using a random value as the context info instead of using it as the salt
parameter is a recommendation from
https://soatok.blog/2021/11/17/understanding-hkdf/