fix: replace AES-CBC with an AEAD mode: AES-CCM
Closes #5705

AES-CBC is malleable and its use is dangereous without using a
encrypt-then-mac approach.
AES-CCM is an AEAD that is widely available and free to use (as opposed
to OCB which was patented until 2021).
AES-CCM is also stronger against nonce misuse than GCM. Until #5706 is
fixed, nonce misuses are possible, so CCM seems preferable, even if it
is slower (2-passes).