Fix locales CDN publish (#5882)
**Problem**  

- `@uppy/locales` on npm used to have `dist` (minified), `lib` (JS),
`src` (to make the sourcemaps work) but dist is redundant tbh and only
for the CDN
- Since the new release tooling only `lib` was published to npm
- CI release could not find any files to upload to CDN (it uses
npm-pack, which checks the "files" in package.json and `dist` was
missing)

**Solution**:

- Correct "files" in package.json's
- Remove `main` from @uppy/locales package.json which pointed to
something random?

We still publish `dist` for `@uppy/locales` needlessly to npm afaik.
Would require a bit more hacking to separate in the current setup