move clear files logic
into #restore instead.
we currently clear files when state transitions to all files complete,
however there's an issue with that where if progress events come in after all files are marked as completed, it will overwrite the metadataStore,
meaning the files that have been cleared will be re-added after they were cleared. this causes files to be restored (when e.g. refreshing the browser)
when they should not (because they have already completed). i managed to reproduce this with the google drive picker plugin (but not with google drive non-picker)